requirements.txt
flask
flask_cors
waitress
google-cloud-firestore
google-cloud-pubsub

# fastapi
# pydantic
